requirements.txt
pyrofork
tgcrypto
motor
fastapi
uvicorn
aiohttp
pytz
python-dotenv
jinja2
psutil